+++ tests/test_overlay_freesurfermesh.py
@@ -5,8 +5,12 @@
 # Author: Paul McCarthy <pauldmccarthy@gmail.com>
 #
 
+import os.path as op
 
 import pytest
+
+import nibabel as nib
+import fsl.data.image as fslimage
 
 from . import run_cli_tests
 
@@ -25,10 +29,24 @@ freesurfer/lh.pial -mc 1 0 0 -o -w 5
 freesurfer/lh.pial -mc 1 0 0 -o -w 10
 freesurfer/lh.pial -mc 1 0 0 -o -w 10 -cm hot -vd freesurfer/lh.curv
 freesurfer/lh.pial -mc 1 0 0 -o -w 10 -cm hot -vd freesurfer/lh.curv
+
+{{asmgh('3d.nii.gz')}} freesurfer/lh.pial -mc 1 0 0 -r 3d.mgh -s torig
+{{asmgh('3d.nii.gz')}} freesurfer/lh.pial -mc 1 0 0 -r 3d.mgh -s affine
+{{asmgh('3d.nii.gz')}} freesurfer/lh.pial -mc 1 0 0 -r 3d.mgh -s pixdim
+{{asmgh('3d.nii.gz')}} freesurfer/lh.pial -mc 1 0 0 -r 3d.mgh -s id
 """
 
+def asmgh(infile):
+    outfile = op.basename(fslimage.removeExt(infile))
+    outfile = outfile + '.mgh'
+    inimg   = fslimage.Image(infile)
+    outimg  = nib.MGHImage(inimg.data, inimg.voxToWorldMat)
+    outimg.to_filename(outfile)
+    return outfile
+
+
 def test_overlay_freesurfermesh():
-    extras = {}
+    extras = {'asmgh' : asmgh}
     run_cli_tests('test_overlay_freesurfermesh',
                   cli_tests,
                   extras=extras,