+++ CHANGELOG.rst
@@ -9,8 +9,8 @@ This document contains the ``fsleyes`` release history in reverse
 chronological order.
 
 
-0.30.0 (Under development)
---------------------------
+0.30.0 (Thursday 27th June 2019)
+--------------------------------
 
 
 Added
@@ -25,6 +25,8 @@ Added
 * New alternative main function :func:`fsleyes.filtermain.main`, which filters
   out useless warnings that originate from underlying libraries (e.g. ``wx``,
   GTK, Cocoa, etc).
+* New ``--cliserver`` option, allowing command-line arguments to be passed
+  to a single FSLeyes instance.
 
 
 Changed
@@ -39,6 +41,8 @@ Changed
 * 2D NIfTI images are now displayed with a 2D texture - this means that
   the maximum dimension size for 2D images is now 16384 on typical
   hardware.
+* The *Freesurfer coordinates* coordinate space option is no longer available
+  on non-freesurfer mesh overlays.
 
 
 0.29.0 (Sunday May 12th 2019)